+++ src/fieldlayoutelements/Template.php
@@ -37,7 +37,7 @@ class Template extends BaseUiElement
     /**
      * @inheritdoc
      */
-    protected function selectorIcon()
+    protected function selectorIcon(): ?string
     {
         return '@appicons/template.svg';
     }