+++ CHANGELOG.md
@@ -100,6 +100,7 @@
 - Fields’ `setIsFresh()` methods must now have a `void` return type declaration.
 - Field layout elements’ `formHtml()` methods must now have a `?string` return type declaration.
 - Field layout elements’ `settingsHtml()` methods must now have a `?string` return type declaration.
+- Field layout UI elements’ `selectorIcon()` methods must now have a `?string` return type declaration.
 - Plugins’ `afterSaveSettings()` methods must now have a `void` return type declaration.
 - Plugins’ `createSettingsModel()` methods must now have a `?Model` return type declaration.
 - Plugins’ `getCpNavItem()` methods must now have an `?array` return type declaration.